BroadleafCommerce/QA-4031 Add default value to Content Items
