+++ admin/broadleaf-contentmanagement-module/src/main/java/org/broadleafcommerce/cms/structure/domain/StructuredContentFieldImpl.java
@@ -17,7 +17,6 @@
  */
 package org.broadleafcommerce.cms.structure.domain;
 
-import org.apache.commons.lang.StringUtils;
 import org.broadleafcommerce.common.copy.CreateResponse;
 import org.broadleafcommerce.common.copy.MultiTenantCopyContext;
 import org.broadleafcommerce.common.extensibility.jpa.copy.DirectCopyTransform;
@@ -76,9 +75,7 @@ public class StructuredContentFieldImpl implements StructuredContentField, Profi
     @Column (name = "VALUE")
     protected String stringValue;
 
-    @AdminPresentation
-    @Column (name = "DEFAULT_VALUE")
-    protected String defaultValue;
+
 
     @AdminPresentation
     @Column (name = "LOB_VALUE", length = Integer.MAX_VALUE - 1)
@@ -110,10 +107,7 @@ public class StructuredContentFieldImpl implements StructuredContentField, Profi
     public String getValue() {
         if (stringValue != null && stringValue.length() > 0) {
             return stringValue;
-        }if (StringUtils.isEmpty(stringValue)&&StringUtils.isEmpty(lobValue)){
-            return defaultValue;
-        }
-        else {
+        } else {
             return lobValue;
         }
     }
@@ -132,17 +126,6 @@ public class StructuredContentFieldImpl implements StructuredContentField, Profi
             lobValue = null;
             stringValue = null;
         }
-    }
-
-    @Override
-    public String getDefaultValue() {
-        return defaultValue;
-    }
-
-
-    @Override
-    public void setDefaultValue(String defaultValue) {
-        this.defaultValue = defaultValue;
     }
 
     @Override
